services/math_service/pyproject.toml
[project]
name = "math_service"
version = "0.2.0"
description = "Math gRPC service"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.66",
    "grpcio-tools>=1.76.0",
    "protobuf>=5.27",
    "shared_protos",
]

[project.optional-dependencies]
dev = ["grpcio-tools>=1.66"]   # 코드 생성용(개발시)

# uv에게 'shared_protos'는 워크스페이스에서 가져오라고 지정
[tool.uv.sources]
shared_protos  = { workspace = true }